+++ src/main/kotlin/ukidelly/api/v2/feed/FeedRoutes.kt
@@ -4,13 +4,7 @@ import io.ktor.resources.*
 
 
 @Resource("/feed")
-class FeedRoutes {
-
-    @Resource("latest")
-    class Latest(val parent: FeedRoutes = FeedRoutes(), val page: Int = 1, val size: Int = 10)
-
-    @Resource("new")
-    class New(val parent: FeedRoutes = FeedRoutes())
+class FeedRoutes(val page: Int = 1, val size: Int = 10) {
 
     @Resource("{feedId}")
     class FeedId(val parent: FeedRoutes = FeedRoutes(), val feedId: Int) {
