+++ src/main/kotlin/ukidelly/api/v2/trade-feed/TradeFeedRouting.kt
@@ -21,7 +21,7 @@ fun Route.tradeFeedRouting() {
 
 
     // 최신 게시글 가져오기
-    get<TradeFeedRoutes.Latest> { feed ->
+    get<TradeFeedRoutes> { feed ->
         val feeds = tradeFeedService.getLatestPosts(feed.size, feed.page)
         call.respond(HttpStatusCode.OK, ResponseDto.Success(feeds, "성공"))
     }
@@ -29,6 +29,7 @@ fun Route.tradeFeedRouting() {
 
     // 게시글 가져오기
     get<TradeFeedRoutes.FeedId> { feed ->
+        
         val feedData = tradeFeedService.getPost(feed.feed_id)
         call.respond(HttpStatusCode.OK, ResponseDto.Success(feedData, "성공"))
     }
@@ -36,7 +37,7 @@ fun Route.tradeFeedRouting() {
     authenticate("auth-jwt") {
 
         // 새 게시물
-        post<TradeFeedRoutes.New> {
+        post<TradeFeedRoutes> {
             val uuid = call.principal<UserIdPrincipal>()!!.name
             val reqeust = call.receive<CreateTradeFeedRequestDto>()
             val newPost = tradeFeedService.addNewPost(reqeust, UUID.fromString(uuid))
